small ui fixes, protecting session-aware server actions

+++ serverActions/NewsHandling/DeleteNews.ts
@@ -1,8 +1,17 @@
 "use server";
 import { AlertMessage } from "./PostNewsAction";
 import { query } from "@/lib/Db";
+import { requireSession } from "@/lib/Auth";
 
 export async function DeleteNews(id: number): Promise<AlertMessage> {
+  const session = await requireSession();
+
+  if (!session) {
+    return {
+      alertType: "error",
+      alertMessage: "Invalid user session",
+    };
+  }
   try {
     await query("DELETE FROM news WHERE id = $1", [id]);
